created local /modules for DonutChart config and abastracted shared functions to /src/modules
